修改了选取代理的方式，无论是在本机器上选取，还是在所有代理中随机选取，一旦发现某个代理不可用就将其从候选队列中删除
